\name{BotanizeR_getlocallist}
\alias{BotanizeR_getlocallist}
\title{Extract a species list from GBIF portal}
\description{
Using coordinates and a buffer radius, this function retrieves a certain species list from GBIF database.
}
\usage{
BotanizeR_getlocallist()
}
\arguments{
\item{long}{a number, setting the longitude.}

\item{lat}{a number, setting the latitude.}

\item{radius}{a number, setting the diameter of the buffer.}

\item{backbone_list}{}
}
\details{
}
\value{
A species list.
}
\details{}
\section{Note}{}

\examples{
# Example
