% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BotanizeR_quiz.R
\name{BotanizeR_quiz}
\alias{BotanizeR_quiz}
\title{BotanizeR quiz}
\usage{
BotanizeR_quiz(
  species_list,
  image_floraweb = TRUE,
  hints_floraweb = c("description", "status", "habitat", "family", "German name"),
  image_ukplantatlas = FALSE,
  hints_ukplantatlas = NULL,
  imagelinks_custom = NULL,
  image_folders = NULL,
  hints_custom = NULL,
  case_sensitive = TRUE,
  file_location = "temporary",
  startat = 0,
  init_count = sum(species_list$COUNT),
  init_score = sum(species_list$SCORE),
  init_attempts = sum(species_list$ATTEMPTS),
  max_attempts = 10,
  image_width = 500
)
}
\arguments{
\item{species_list}{a data.frame with the species for which we want to
retrieve information. **It should contain the following columns**: 
*NAMNR*, *TAXONNAME*, *SPECIES*, *GENUS*, *EPITHET*, *AUTHOR*, *COUNT*,
*SCORE* and *ATTEMPTS*}

\item{image_floraweb}{logical that defines if images from
[FloraWeb](https://www.floraweb.de) should be retrieved}

\item{hints_floraweb}{character vector that defines what hints the user
wants to retrieve from [FloraWeb](https://www.floraweb.de)}

\item{image_ukplantatlas}{logical that defines if images from the
[Online Atlas of the British and Irish flora ](https://www.brc.ac.uk/plantatlas/)
should be retrieved}

\item{hints_ukplantatlas}{character vector that defines what hints the user
wants to retrieve from the
[Online Atlas of the British and Irish flora ](https://www.brc.ac.uk/plantatlas/)}

\item{imagelinks_custom}{character vector that defines a custom link to
retrieve images}

\item{image_folders}{character vector that defines a specific folder from
which the user wants to retrieve images}

\item{hints_custom}{character vector that defines personal hints the user
wants to use. **Note:** in that case, these hints should be present as
columns in the `species_row` table.}

\item{case_sensitive}{logical.}

\item{file_location}{character vector that defines}

\item{startat}{number}

\item{init_count}{sum of the column `$COUNT` in `species_list`,
defining the total number of tries of the user}

\item{init_score}{sum of the column `$SCORE` in `species_list`,
defining the total score of the user}

\item{init_attempts}{sum of the column `$ATTEMPTS` in `species_list`,
defining how many attempts he user had in total}

\item{max_attempts}{number defining the number of attempts per species}

\item{image_width}{number to define the width of the images.}
}
\value{
A data.frame with updated scores and counts per species.
}
\description{
Navigates through species from floraweb.de (or a defined subset thereof),
shows pictures and descriptions and other hints and let's the user guess the
species name. Number of tries and attempts are used to calculate scores that
will be used to update probabilities for the random sampling of subsequent
species.
}
\details{
After running the main function BotanizeR_quiz() you need to click
into the console to type the species names in there. If you have no clue,
press enter and the next hint will appear (you have ten tries). If you type
the name slightly wrong the function will tell you. If the genus is correct
it will also tell you. It shows several photos if available. If you want to
skip a species write "skip". If you want to cancel the quiz write "exit".
Don't hit Esc if you want to save your progress. The function counts your
attempts and successes and uses this as sampling probabilities for choosing
the species randomly. The better you know a species the less likely it will
be shown again.
}
\examples{
# Species list for Germany with IDs from floraweb.de
data(floraweb_species)

# Subset for about 300 species students in Goettingen learn
floraweb_species <- floraweb_species[which(floraweb_species$SUMMER==1 |
floraweb_species$BioDiv2005==1), ]

# Type in species name, or press enter for next hint or type "skip" and press 
# enter for next species or type "exit" to end quiz and save results
floraweb_species_trained <- BotanizeR_quiz(
species_list = floraweb_species, hints_floraweb = c("description", "status",
"habitat","family","German name"), case_sensitive = FALSE)

# If you want to include distribution maps as hints add "map" to hints; This increases the download times a bit
floraweb_species_trained <- BotanizeR_quiz(
species_list = floraweb_species,
hints_floraweb = c("map","description","status", "habitat", "family",
"German name"), case_sensitive = FALSE)

# If you want to keep track of your progress, you can save the species list with 
# updated scores locally and load it in the next session

# Initial saving
write.csv(floraweb_species_trained, "floraweb_species_trained.csv",
row.names = FALSE)

# Load species list
floraweb_species_trained <- read.csv("floraweb_species_trained.csv")

# Practice
floraweb_species_trained <- BotanizeR_quiz(
  species_list = floraweb_species_trained,
  hints_floraweb = c("map", "description", "status", "habitat", "family",
                     "German name"), case_sensitive = FALSE)

# Save species list
write.csv(floraweb_species_trained, "floraweb_species_trained.csv",
          row.names = FALSE)

### example for three species with custom hints and images
custom_species <- floraweb_species[
  which(floraweb_species$SPECIES \%in\% c("Acer campestre", "Erica carnea",
                                        "Melampyrum nemorosum")), ]

custom_species_trained <- BotanizeR_quiz(
  species_list = custom_species, image_floraweb = TRUE,
  hints_floraweb = NULL, hints_custom = c("ownhint_1", "ownhint_2"), 
  imagelinks_custom = c("imagelink_1", "imagelink_2"), 
  image_folders = NULL, case_sensitive = FALSE)

}
\references{
Weigelt, P., Denelle, P., Brambach, F. & Kreft, H. (2021) A flexible
    R-package with Shiny-App for practicing plant identification in times of
    online teaching and beyond. Plants, People, Planet.
}
\seealso{
[BotanizeR::BotanizeR_collect()]
}
