% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BotanizeR_shiny.R
\name{BotanizeR_shiny}
\alias{BotanizeR_shiny}
\title{BotanizeR Shiny application}
\usage{
BotanizeR_shiny(run = TRUE)
}
\arguments{
\item{run}{a logical, when TRUE, the Shiny application is launched. Can be
set to FALSE to avoid launching the shiny app in a non-interactive
environment.}
}
\value{
Shiny application.
}
\description{
Running the Shiny application of BotanizeR locally from installed BotanizeR
R package.
}
\examples{
# Example
\dontrun{
BotanizeR_shiny()
}

}
\references{
\preformatted{Weigelt, P., Denelle, P., Brambach, F. & Kreft, H. (2021) A flexible
R package with Shiny app to practice plant identification for 
online teaching and beyond. PLANTS, PEOPLE, PLANET, 
https://doi.org/10.1002/ppp3.10226.
}
}
