% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BotanizeR_species.R
\docType{data}
\name{BotanizeR_species}
\alias{BotanizeR_species}
\title{Species table for BotanizeR}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://www.floraweb.de}{FloraWeb}
}
\usage{
data("BotanizeR_species")
}
\description{
Table containing the relevant information for 5852 species. The data
originate from \href{https://www.floraweb.de}{FloraWeb}.
}
\examples{
data(BotanizeR_species)
}
\keyword{datasets}
