% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGRS_Germany.R
\docType{data}
\name{CGRS_Germany}
\alias{CGRS_Germany}
\title{Species table for BotanizeR}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://www.eea.europa.eu/data-and-maps/data/common-european-chorological-grid-reference-system-cgrs}{Common European Chorological Grid Reference System (CGRS)}
}
\usage{
data("CGRS_Germany")
}
\description{
Chorological Grid Reference System for Germany.
}
\examples{
data(CGRS_Germany)
}
\keyword{datasets}
