% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floraweb_species.R
\docType{data}
\name{floraweb_species}
\alias{floraweb_species}
\title{Species table for BotanizeR}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://www.floraweb.de}{FloraWeb}
}
\usage{
data("floraweb_species")
}
\description{
Table containing the relevant information for 4639 species. The data
originate from [FloraWeb](https://www.floraweb.de).
}
\examples{
data(floraweb_species)
}
\keyword{datasets}
