% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukplantatlas_species.R
\docType{data}
\name{ukplantatlas_species}
\alias{ukplantatlas_species}
\title{Species table for BotanizeR}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://www.brc.ac.uk/plantatlas/}{Online Atlas of the British and Irish flora}
}
\usage{
data("ukplantatlas_species")
}
\description{
Table containing the relevant information for 2945 species. The data
originate from [Online Atlas of the British and Irish flora](https://www.brc.ac.uk/plantatlas/).
}
\examples{
data(ukplantatlas_species)
}
\keyword{datasets}
