\name{BotanizeR_collect}
\alias{BotanizeR_collect}
\title{Collect information to show as hints in BotanizeR quiz}
\description{
Collects information from floraweb.de (images, map and species desciptions) and from user defined image folders and columns in the species_list data.frame to show them as hints in BotanizeR_quiz
}
\usage{
BotanizeR_collect()
}
\arguments{
}
\details{
}
\value{
A vector of new probabilities per species.
}
\details{}
\section{Note}{}

\examples{
}
