% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BotanizeR_getlocallist.R
\name{BotanizeR_getlocallist}
\alias{BotanizeR_getlocallist}
\title{GBIF species list}
\usage{
BotanizeR_getlocallist(long = NA, lat = NA, radius = 1, backbone_list = NA)
}
\arguments{
\item{long}{a numeric, defining the longitude of the wanted location.}

\item{lat}{a numeric, defining the latitude of the wanted location.}

\item{radius}{a numeric, defining the radius of the buffer around the
wanted location.}

\item{backbone_list}{a data.frame with the same columns as the data
returned by data("BotanizeR_species").}
}
\value{
A species list.
}
\description{
Using coordinates and a buffer radius, this function retrieves a certain
species list from \href{https://www.gbif.org/}{GBIF} database.
}
\examples{
# Example

}
\references{
\preformatted{Weigelt, P., Denelle, P., Brambach, F. & Kreft, H. (2021) A flexible
R-package with Shiny-App for practicing plant identification in times of
online teaching and beyond. submitted.
}
}
