% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BotanizeR_imageresize.R
\name{BotanizeR_imageresize}
\alias{BotanizeR_imageresize}
\title{Resizing images}
\usage{
BotanizeR_imageresize(
  image_folders = NULL,
  image_width = NA,
  max_height = NA,
  int_type = 1,
  quality = 1
)
}
\arguments{
\item{image_folders}{character vector that defines a specific folder from
which the user wants to retrieve images}

\item{image_width}{number to define the width of the images.}

\item{max_height}{number to define the height of the images.}

\item{int_type}{number to define the type of the images.}

\item{quality}{number to define the quality of the images.}
}
\value{
Resized images.
}
\description{
Batch resize images for plotting in BotanizeR quiz.
}
\details{
Resizing images so they fit in the Shiny application.
}
\examples{
# Example

}
\references{
\preformatted{Weigelt, P., Denelle, P., Brambach, F. & Kreft, H. (2021) A flexible
R-package with Shiny-App for practicing plant identification in times of
online teaching and beyond. Plants, People, Planet.
}
}
\seealso{
\code{\link[imager:resize]{imager::resize()}}
}
