\name{BotanizeR_shiny}
\alias{BotanizeR_shiny}
\title{Run BotanizeR Shiny app in R console}
\description{
Runs locally the Shiny application associated to BotanizeR package.
}
\usage{
BotanizeR_shiny()
}
\arguments{
\item{run}{a binary, runs the Shiny application when set to TRUE.}
}
\details{
}
\value{
}
\details{}
\section{Note}{}

\examples{
BotanizeR_shiny(run = FALSE)
