% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BotanizeR_shiny.R
\name{BotanizeR_shiny}
\alias{BotanizeR_shiny}
\title{Shiny application}
\usage{
BotanizeR_shiny(run = FALSE)
}
\arguments{
\item{run}{a logical, when TRUE, the Shiny application is lauched}
}
\value{
Shiny application.
}
\description{
Running the Shiny application of BotanizeR.
}
\examples{
# Example
BotanizeR_shiny(run = FALSE)

}
\references{
\preformatted{Weigelt, P., Denelle, P., Brambach, F. & Kreft, H. (2021) A flexible
R-package with Shiny-App for practicing plant identification in times of
online teaching and beyond. submitted.
}
}
