% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BotanizeR_species.R
\docType{data}
\name{BotanizeR_species}
\alias{BotanizeR_species}
\title{Species table for BotanizeR}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://www.floraweb.de}{FloraWeb} and
\href{https://www.brc.ac.uk/plantatlas/}{Online Atlas of the
British Irish flora}
}
\usage{
data("BotanizeR_species")
}
\description{
Example species list table containing the relevant information to run
\code{\link[=BotanizeR_quiz]{BotanizeR_quiz()}} and \code{\link[=BotanizeR_shiny]{BotanizeR_shiny()}} for 5852
species from \href{https://www.floraweb.de}{FloraWeb}
and the \href{https://www.brc.ac.uk/plantatlas/}{Online Atlas of the British Irish flora}.
}
\examples{
data(BotanizeR_species)
}
\keyword{datasets}
